clean: utils関数を各ファイルに切り出すように整理、インポートの修正
