+++ components.json
@@ -12,7 +12,7 @@
   },
   "aliases": {
     "components": "@/components",
-    "utils": "@/lib/utils",
+    "utils": "@/utils/common",
     "ui": "@/components/ui",
     "lib": "@/lib",
     "hooks": "@/hooks"
