+++ utils/common.ts
@@ -1,0 +1,10 @@
+import { clsx, type ClassValue } from "clsx";
+import { twMerge } from "tailwind-merge";
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
+
+export const formatTagForLink = (tag: string) => {
+  return tag.toLowerCase().replaceAll(" ", "");
+};
